+++ tsl/src/compression/algorithms/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/array.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/datum_serialize.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/deltadelta.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/dictionary.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/gorilla.c)
+target_sources(${TSL_LIBRARY_NAME} PRIVATE ${SOURCES})
