+++ tsl/src/compression/algorithms/array.c
@@ -14,11 +14,11 @@
 #include <utils/lsyscache.h>
 #include <utils/syscache.h>
 
-#include "compression/array.h"
+#include "array.h"
 #include "compression/compression.h"
-#include "compression/simple8b_rle.h"
-#include "compression/simple8b_rle_bitmap.h"
 #include "datum_serialize.h"
+#include "simple8b_rle.h"
+#include "simple8b_rle_bitmap.h"
 
 #include "compression/arrow_c_data_interface.h"
