+++ tsl/src/compression/algorithms/gorilla.c
@@ -14,14 +14,14 @@
 #include <utils/builtins.h>
 #include <utils/memutils.h>
 
-#include "compression/gorilla.h"
+#include "gorilla.h"
 
 #include "adts/bit_array.h"
 #include "compression/arrow_c_data_interface.h"
 #include "compression/compression.h"
-#include "compression/simple8b_rle.h"
-#include "compression/simple8b_rle_bitmap.h"
 #include "float_utils.h"
+#include "simple8b_rle.h"
+#include "simple8b_rle_bitmap.h"
 
 /*
  * Gorilla compressed data is stored as
