+++ tsl/src/init.c
@@ -16,13 +16,13 @@
 #include "bgw_policy/retention_api.h"
 #include "chunk.h"
 #include "chunk_api.h"
+#include "compression/algorithms/array.h"
+#include "compression/algorithms/deltadelta.h"
+#include "compression/algorithms/dictionary.h"
+#include "compression/algorithms/gorilla.h"
 #include "compression/api.h"
-#include "compression/array.h"
 #include "compression/compression.h"
 #include "compression/create.h"
-#include "compression/deltadelta.h"
-#include "compression/dictionary.h"
-#include "compression/gorilla.h"
 #include "compression/segment_meta.h"
 #include "config.h"
 #include "continuous_aggs/create.h"
