+++ tsl/src/nodes/decompress_chunk/exec.c
@@ -21,7 +21,6 @@
 #include <tcop/tcopprot.h>
 
 #include "compat/compat.h"
-#include "compression/array.h"
 #include "compression/arrow_c_data_interface.h"
 #include "compression/compression.h"
 #include "guc.h"
